code refactoring, adding logic for pressing enter in search, props/interfaces refinements.
