+++ ui/frontend/src/core/types/paginationTypes.ts
@@ -1,0 +1,5 @@
+export interface IPaginationProps {
+  total: number
+  page: number
+  onChange: (page: number) => void
+}
