+++ ui/frontend/src/core/types/searchTypes.ts
@@ -1,0 +1,3 @@
+export interface ISearchProps {
+  isLoading: boolean
+}
