+++ ui/frontend/src/core/utils/extractFunctions.ts
@@ -1,0 +1,9 @@
+export const extractUsername = (input: string): string => {
+  const match = input.match(/^@([^ ]+)/);
+  return match ? match[1] : '';
+}
+
+export const extractFullName = (input: string): string => {
+  const match = input.match(/\(([^)]+)\)/);
+  return match ? match[1] : '';
+}
